new_requirements.txt
alembic==1.5.5
bcrypt==3.1.7
blinker==1.4
cachetools==3.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
configobj==5.0.6
dnspython==1.16.0
email-validator==1.1.2
enum34==1.1.10
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
flask-ngrok==0.0.25
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
google-auth==1.28.0
google-auth-oauthlib==0.4.1
gspread==3.7.0
gunicorn==19.10.0
httplib2==0.19.0
idna==2.10
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-editor==1.0.4
python2-secrets==1.0.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.5
serial==0.0.97
six==1.15.0
SQLAlchemy==1.3.23
systematic==4.8.7
typing==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
